properties: refactor two_way_binding code into its own module
